+++ bindings/python/examples/client/02_generate_addresses.py
@@ -18,13 +18,13 @@ secret_manager = MnemonicSecretManager(
     os.environ['NON_SECURE_USE_OF_DEVELOPMENT_MNEMONIC_1'])
 
 # Generate public address with default account index and range.
-addresses = client.generate_addresses(secret_manager)
+addresses = client.generate_ed25519_addresses(secret_manager)
 
 print('List of generated public addresses:', *addresses, sep='\n')
 print()
 
 # Generate public address with custom account index and range.
-addresses = client.generate_addresses(secret_manager,
+addresses = client.generate_ed25519_addresses(secret_manager,
                                       account_index=0,
                                       start=0,
                                       end=4)
@@ -33,7 +33,7 @@ print('List of generated public addresses:', *addresses, sep='\n')
 print()
 
 # Generate internal addresses with custom account index and range.
-addresses = client.generate_addresses(secret_manager,
+addresses = client.generate_ed25519_addresses(secret_manager,
                                       account_index=0,
                                       start=0,
                                       end=4,
@@ -43,13 +43,13 @@ print('List of generated internal addresses:', *addresses, sep='\n')
 print()
 
 # Generate addresses with providing all inputs, that way it can also be done offline without a node.
-addresses = client.generate_addresses(secret_manager,
+addresses = client.generate_ed25519_addresses(secret_manager,
                                       coin_type=CoinType.SHIMMER,
                                       account_index=0,
                                       start=0,
                                       end=4,
                                       internal=False,
-                                      # Generating addresses with client.generateAddresses(secretManager, options={}), will by default get the bech32_hrp (Bech32
+                                      # Generating addresses with client.generateEd25519Addresses(secretManager, options={}), will by default get the bech32_hrp (Bech32
                                       # human readable part) from the node info, generating it "offline" requires setting it in the generateAddressesOptions
                                       bech32_hrp='rms')