+++ sdk/src/wallet/account/operations/syncing/addresses/output_ids/alias_foundry.rs
@@ -11,8 +11,9 @@ use crate::{
     types::{
         api::plugins::indexer::OutputIdsResponse,
         block::{
-            address::{Address, AliasAddress, Bech32AddressLike},
+            address::{AliasAddress, Bech32Address, ToBech32Ext},
             output::{Output, OutputId},
+            ConvertTo,
         },
     },
     wallet::{
@@ -25,12 +26,12 @@ impl Account {
     /// Returns output ids of alias outputs
     pub(crate) async fn get_alias_and_foundry_output_ids(
         &self,
-        bech32_address: impl Bech32AddressLike,
+        bech32_address: impl ConvertTo<Bech32Address>,
         sync_options: &SyncOptions,
     ) -> crate::wallet::Result<Vec<OutputId>> {
         log::debug!("[SYNC] get_alias_and_foundry_output_ids");
         let client = self.client();
-        let bech32_address = bech32_address.to_bech32()?;
+        let bech32_address = bech32_address.convert()?;
 
         let mut output_ids = HashSet::new();
 
@@ -113,7 +114,7 @@ impl Account {
             if let Output::Alias(alias_output) = alias_output_with_meta.output() {
                 let alias_address =
                     AliasAddress::from(alias_output.alias_id_non_null(alias_output_with_meta.metadata().output_id()));
-                let alias_bech32_address = Address::Alias(alias_address).to_bech32(bech32_hrp);
+                let alias_bech32_address = alias_address.to_bech32(bech32_hrp);
                 let client = self.client().clone();
                 tasks.push(Box::pin(task::spawn(async move {
                     client