+++ sdk/src/wallet/bindings/nodejs/types/address.ts
@@ -1,4 +1,5 @@
 import type { HexEncodedAmount } from '@iota/types';
+import type { CoinType } from './account';
 
 /** Address Types */
 export enum AddressType {
@@ -49,4 +50,24 @@ export interface SendNftParams {
 export interface GenerateAddressOptions {
     internal: boolean;
     ledgerNanoPrompt: boolean;
+}
+
+export interface GenerateAddressesOptions {
+    coinType?: CoinType;
+    accountIndex?: number;
+    range?: IRange;
+    /**
+     * Internal addresses
+     */
+    internal?: boolean;
+    /**
+     * Bech32 human readable part
+     */
+    bech32Hrp?: string;
+    options?: GenerateAddressOptions;
+}
+
+export interface IRange {
+    start: number;
+    end: number;
 }