+++ sdk/src/wallet/message_interface/message_handler.rs
@@ -23,12 +23,12 @@ use crate::{
         request_funds_from_faucet, utils, Client, NodeInfoWrapper,
     },
     types::block::{
-        address::HrpLike,
+        address::{Hrp, ToBech32Ext},
         output::{
             dto::{OutputBuilderAmountDto, OutputDto},
             AliasOutput, BasicOutput, FoundryOutput, NativeToken, NftOutput, Output, Rent,
         },
-        Error,
+        ConvertTo, Error,
     },
     wallet::{
         account::{
@@ -256,7 +256,7 @@ impl WalletMessageHandler {
                 })
                 .await
             }
-            Message::GenerateAddress {
+            Message::GenerateEd25519Address {
                 account_index,
                 address_index,
                 options,
@@ -265,7 +265,7 @@ impl WalletMessageHandler {
                 convert_async_panics(|| async {
                     let address = self
                         .wallet
-                        .generate_address(account_index, address_index, options)
+                        .generate_ed25519_address(account_index, address_index, options)
                         .await?;
 
                     let bech32_hrp = match bech32_hrp {
@@ -571,9 +571,19 @@ impl WalletMessageHandler {
                 })
                 .await
             }
-            AccountMethod::GenerateAddresses { amount, options } => {
-                let address = account.generate_addresses(amount, options).await?;
-                Ok(Response::GeneratedAddress(address))
+            AccountMethod::GenerateEd25519Addresses { amount, options } => {
+                let address = account.generate_ed25519_addresses(amount, options).await?;
+                Ok(Response::GeneratedEd25519Addresses(address))
+            }
+            AccountMethod::GenerateEvmAddresses { options } => {
+                let addresses = account
+                    .wallet
+                    .secret_manager
+                    .read()
+                    .await
+                    .generate_evm_addresses(options)
+                    .await?;
+                Ok(Response::GeneratedEvmAddresses(addresses))
             }
             AccountMethod::GetOutputsWithAdditionalUnlockConditions { outputs_to_claim } => {
                 let output_ids = account
@@ -992,7 +1002,7 @@ impl WalletMessageHandler {
     }
 
     /// The create account message handler.
-    async fn create_account(&self, alias: Option<String>, bech32_hrp: Option<impl HrpLike>) -> Result<Response> {
+    async fn create_account(&self, alias: Option<String>, bech32_hrp: Option<impl ConvertTo<Hrp>>) -> Result<Response> {
         let mut builder = self.wallet.create_account();
 
         if let Some(alias) = alias {
@@ -1000,7 +1010,7 @@ impl WalletMessageHandler {
         }
 
         if let Some(bech32_hrp) = bech32_hrp {
-            builder = builder.with_bech32_hrp(bech32_hrp.to_hrp()?);
+            builder = builder.with_bech32_hrp(bech32_hrp.convert()?);
         }
 
         match builder.finish().await {