+++ sdk/tests/wallet/claim_outputs.rs
@@ -29,8 +29,8 @@ async fn claim_2_basic_micro_outputs() -> Result<()> {
     let tx = accounts[1]
         .send_amount(
             vec![
-                SendAmountParams::new(*accounts[0].addresses().await?[0].address(), micro_amount),
-                SendAmountParams::new(*accounts[0].addresses().await?[0].address(), micro_amount),
+                SendAmountParams::new(*accounts[0].addresses().await?[0].address(), micro_amount)?,
+                SendAmountParams::new(*accounts[0].addresses().await?[0].address(), micro_amount)?,
             ],
             TransactionOptions {
                 allow_micro_amount: true,
@@ -83,8 +83,8 @@ async fn claim_1_of_2_basic_outputs() -> Result<()> {
     let tx = accounts[1]
         .send_amount(
             vec![
-                SendAmountParams::new(*accounts[0].addresses().await?[0].address(), amount),
-                SendAmountParams::new(*accounts[0].addresses().await?[0].address(), 0),
+                SendAmountParams::new(*accounts[0].addresses().await?[0].address(), amount)?,
+                SendAmountParams::new(*accounts[0].addresses().await?[0].address(), 0)?,
             ],
             TransactionOptions {
                 allow_micro_amount: true,
@@ -236,18 +236,14 @@ async fn claim_2_native_tokens() -> Result<()> {
     let tx = accounts[1]
         .send_native_tokens(
             vec![
-                SendNativeTokensParams {
-                    address: *accounts[0].addresses().await?[0].address(),
-                    native_tokens: vec![(mint_tx_0.token_id, native_token_amount)],
-                    expiration: None,
-                    return_address: None,
-                },
-                SendNativeTokensParams {
-                    address: *accounts[0].addresses().await?[0].address(),
-                    native_tokens: vec![(mint_tx_1.token_id, native_token_amount)],
-                    expiration: None,
-                    return_address: None,
-                },
+                SendNativeTokensParams::new(
+                    *accounts[0].addresses().await?[0].address(),
+                    [(mint_tx_0.token_id, native_token_amount)],
+                )?,
+                SendNativeTokensParams::new(
+                    *accounts[0].addresses().await?[0].address(),
+                    [(mint_tx_1.token_id, native_token_amount)],
+                )?,
             ],
             None,
         )
@@ -534,6 +530,51 @@ async fn claim_2_nft_outputs_no_outputs_in_claim_account() -> Result<()> {
     let balance = account_1.sync(None).await.unwrap();
     assert_eq!(balance.potentially_locked_outputs().len(), 0);
     assert_eq!(balance.nfts().len(), 2);
+
+    tear_down(storage_path)
+}
+
+#[ignore]
+#[tokio::test]
+async fn claim_basic_micro_output_error() -> Result<()> {
+    let storage_path = "test-storage/claim_basic_micro_output_error";
+    setup(storage_path)?;
+
+    let wallet = make_wallet(storage_path, None, None).await?;
+
+    let account_0 = &create_accounts_with_funds(&wallet, 1).await?[0];
+    let account_1 = wallet.create_account().finish().await?;
+
+    let micro_amount = 1;
+    let tx = account_0
+        .send_amount(
+            vec![SendAmountParams::new(
+                *account_1.addresses().await?[0].address(),
+                micro_amount,
+            )?],
+            TransactionOptions {
+                allow_micro_amount: true,
+                ..Default::default()
+            },
+        )
+        .await?;
+
+    account_0
+        .retry_transaction_until_included(&tx.transaction_id, None, None)
+        .await?;
+
+    // Try claim with account 1 will fail since it has no funds to cover the storage deposit
+    let balance = account_1.sync(None).await.unwrap();
+    assert_eq!(balance.potentially_locked_outputs().len(), 1);
+
+    let result = account_1
+        .claim_outputs(
+            account_1
+                .get_unlockable_outputs_with_additional_unlock_conditions(OutputsToClaim::MicroTransactions)
+                .await?,
+        )
+        .await;
+    assert!(matches!(result, Err(iota_sdk::wallet::Error::InsufficientFunds { .. })));
 
     tear_down(storage_path)
 }